move config to xml, got basic graphics working
I move the config file over to xml. Apache has a libary that makes this easy to parse once you've got it set up.

In addition, I copied the ImageFont class over from an old package so that I could get simple graphics up and running. I need to do a lot of refactoring with that class to make it less hard coded and more config dependent. Hard coded == bad.

But the good news is I've got a basic working prototype up that draws a single glyph onto the screen. Next, on to conquer the world!